refactor(settings): extract AccountFormModal (#166)
The add/edit form moves verbatim from SettingsView (2745 -> 801
lines) into components/settings/AccountFormModal.vue on ModalShell:
form state + selectAccountType switch, save/cancel, Google/
Microsoft OAuth, JMAP OIDC device flow, Talk/Matrix/Zoom sign-in,
IMAP autodiscovery, default-book selection, and the form CSS. The
view drives it via exposed openNew(type)/openEdit(id); it keeps
the account lists, prefs, picker wiring, delete-confirm and the
?addAccount deep link (whose provider map, as before, does not
include fastmail — now covered by a test comment). Invariants
kept: defaultForm() before selectAccountType on openNew; edit
never calls selectAccountType; OIDC oauth2: marker only for new
accounts. +4 settings-view tests.
